No shoulds, == to be_, field contains/equals
Changed 'should see' and 'should not see' to just 'see' and 'do not
see'. This is consistent with the principle that each step of a test
case should be an assertion, verifiable as either true or false. Whether
I *should* see something is a question for the user interface designers;
at test time, the important question is whether I *do* see something.
This is a subtle distinction (and goes against the grain of what has
come before, particularly Cucumber / Capybara / Kelp) but I think it
results in more succinct tests. The visibility question during testing
then becomes, "Do I see 'Foo'?" (yes or no). Not "Should I see 'Foo'?"
(in which case, what does a 'no' answer mean?)

Also added text-field content verification, and changed "== true" to
"be_true" and so forth in the spec tests.