Highlight located element, fix checkboxes
The checkbox 'check' and 'uncheck' methods do not appear to cause
'onclick' events to be emitted; hence, any checkbox that kicks off some
Javascript code on checkbox enable/disable was not getting executed.
This commit modifies the (enable|disable)_checkbox methods to use the
'click' method instead, and skip clicking if the checkbox is already in
the desired state.

Also added highlight_located_element = true, to make it easier to see
what Selenium is doing while tests are running.