Got 'within' scoping to work for buttons(?)
The prior commit included two failing spec tests related to finding
buttons within a scope. As anything with XPaths, the situation is
complicated, but the gist is that the XPath::HTML.button method returns
a "Union", which basically means an ORed expression that could match any
one of several things (an input with a certain label, an input with an
image, a button with a label, etc.). This is all well and good when
they're used by themselves (with no predicates, like I'm trying to do
with 'within' here), but failed in the context of a scope, because:

    scope.child(union)

effectively only applied the scoping qualifier to the *first* expression
in the union, leaving the remaining expressions in the union "naked"
(without the scope qualifier). In simpler terms, if the scope is "A",
then doing:

    scope.child(B, C, D)

gave:

    AB | C | D

instead of:

    AB | AC | AD

like I wanted it to. The solution? A scary-looking but working recursive
hack that splits the XPath::Union into its component expressions,
applies the scope to each sub-expression, then rejoins them into a union
again. This allows the scope to apply to each expression in the union,
instead of just the first one.

All spec tests are passing once again.