SeleniumTest config as a hash, strip HTML
Overhaul SeleniumTest#initialize so it accepts a hash of configuration
parameters instead of positional arguments. This allows easy setting of
any optional parameter without having to pass *all* parameters. Included
stop_on_error in these optional parameters, so that can be set from a
FitNesse Slim script table now.

Also did quick HTML-stripping of the `url` parameter, which allows the
URL to be unescaped in the FitNesse table. Since an unescaped table is
necessary for hash parameters to work, this should help reduce annoying
failures caused by not remembering to escape the URL.