Fail on disabled form inputs
The Selenium-client methods for editing form widgets (checking
checkboxes, selecting radiobuttons, typing into fields, selecting from
dropdowns) for whatever reason are *successful* when the respective
widget is *disabled* (not user-editable). Whether this is a deliberate
design decision or a legitimate Selenium-client bug is unknown;
regardless, this behavior is not desired in Rsel because it allows Rsel
to do things that would be impossible in a real browser.

This commit changes that behavior, causing the following methods to fail
if the widget is disabled/read-only:

- click_button
- type_into_field
- fill_in_field
- select_radio
- select_from_dropdown
- enable_checkbox
- disable_checkbox

Included spec tests for all of these, all passing now.