Split up selenium_test_spec
This commit splits up the monolithic selenium_test_spec.rb into many
shorter spec files, allowing for easier execution of individual groups
of tests. Doing it this way requires an alternative way of sharing the
same browser instance across example groups; this is accomplished with a
@@st class variable, which is a kind of ugly way to do it, and RSpec
balks about something in the configuration block, but it gets the job
done and allows the tests to still run quickly (~5 minutes for the full
suite on my box).

Also using "require 'spec/spec_helper'" instead of mucking about with
dirnamd and expand_path.