Add some pending specs for hidden fields
From the user's perspective, hidden fields in a form ought to be
non-interactive--that is, there should be no way to fill them in, and no
way to verify their contents. However, due to an apparent bug in
selenium-client, hidden fields are considered editable (is_editable
returns true), and due to a possible bug in the 'tagname' method, fields
of type="hidden" are evaluated as 'input.text', allowing their contents
to be inspected with generic_field_equals.

This commit adds two pending specs for hidden fields, since we might
want to address this matter at some point (though I wouldn't consider it
high-priority).