First attempt at using WebDriver
This branch has a semi-functional implementation of WebDriver-backed
Selenium:

  http://code.google.com/p/selenium/wiki/RubyBindings#-backed_Selenium

This approach is probably not going to give us optimal performance; it
is possible to use straight-up WebDriver with remote calls to the
selenium-server, but even that appears to have pretty dismal
performance:

  http://groups.google.com/group/webdriver/browse_thread/thread/aa162c93061910e8

But considering that the selenium-client gem we've been using is a
couple years old, and appears to be unmaintained, we may have little
choice but to switch to WebDriver eventually.

A lot of Rsel code changes will be needed to adapt to the WebDriver way
of doing things. For instance, where we used to do:

  browser.click("link=Form test")
  browser.type("xpath=.//some/xpath", "Foo")

We would do something like:

  browser.find_element(:link, "About this site").click
  browser.find_element(:xpath, ".//some/xpath").send_keys("Foo")

The selenium-webdriver documentation leaves something to be desired; the
API appears to be distributed through a number of lower-level classes,
but here are some starting points:

  http://rdoc.info/github/lanej/selenium-webdriver/frames
  http://rdoc.info/github/lanej/selenium-webdriver/master/Selenium/WebDriver/Driver
  http://rdoc.info/github/lanej/selenium-webdriver/master/Selenium/WebDriver/Firefox/Bridge

Be sure to run:

  $ bundle install

to ensure that the selenium-webdriver gem gets installed. This gem
provides everything that the old selenium-client gem provided, so we can
probably take selenium-client out of the .gemspec.